// src/app/model/loginResponse.model.ts
export class LoginResponse {

    enforcePassChange: string;
    fullName: string;
    userId: string;
    username: string;
    branchNo: string;
    branchName: string;
    branchCode: string;
    deptId: string;
    bankingDate: string;
    status: string;
    roleId: string;
    roleName: string;
    lastLoginDate: string;
    errorCode: number;
    licenseErrorCode: number;
    licenseNumberOfDay: number;
    errorMessage: string;
    classCode: string;
    url: string;

}